refactor: unifica logger e remove console
