+++ backend/src/services/logger.ts
@@ -1,4 +1,6 @@
 import winston from 'winston';
+import DailyRotateFile from 'winston-daily-rotate-file';
+import path from 'path';
 
 // Configuração de formato personalizado
 const customFormat = winston.format.combine(
@@ -22,9 +24,25 @@ const customFormat = winston.format.combine(
   })
 );
 
-// Configuração dos transportes
+// Diretório de logs e transportes
+const logDir = process.env.LOG_DIR || 'logs';
+
 const transports: winston.transport[] = [
-  // Console transport
+  new DailyRotateFile({
+    filename: path.join(logDir, 'application-%DATE%.log'),
+    datePattern: 'YYYY-MM-DD',
+    maxSize: '20m',
+    maxFiles: '14d',
+    format: customFormat
+  }),
+  new DailyRotateFile({
+    filename: path.join(logDir, 'error-%DATE%.log'),
+    datePattern: 'YYYY-MM-DD',
+    level: 'error',
+    maxSize: '20m',
+    maxFiles: '30d',
+    format: customFormat
+  }),
   new winston.transports.Console({
     level: process.env.NODE_ENV === 'production' ? 'info' : 'debug',
     format: winston.format.combine(
@@ -33,25 +51,6 @@ const transports: winston.transport[] = [
     )
   })
 ];
-
-// Em produção, adicionar arquivo de log
-if (process.env.NODE_ENV === 'production') {
-  transports.push(
-    new winston.transports.File({
-      filename: 'logs/error.log',
-      level: 'error',
-      format: customFormat,
-      maxsize: 5242880, // 5MB
-      maxFiles: 5,
-    }),
-    new winston.transports.File({
-      filename: 'logs/combined.log',
-      format: customFormat,
-      maxsize: 5242880, // 5MB
-      maxFiles: 5,
-    })
-  );
-}
 
 // Criar logger
 const logger = winston.createLogger({
@@ -109,6 +108,52 @@ export const loggerService = {
     });
   }
 };
+
+// Middleware para logs de requisições
+export const requestLogger = (req: any, res: any, next: any) => {
+  const startTime = Date.now();
+
+  res.on('finish', () => {
+    const duration = Date.now() - startTime;
+    logger.info('HTTP Request', {
+      method: req.method,
+      url: req.url,
+      statusCode: res.statusCode,
+      duration: `${duration}ms`,
+      ip: req.ip,
+      userAgent: req.get('User-Agent')
+    });
+  });
+
+  next();
+};
+
+// Função para log de erros com contexto
+export const logError = (error: Error, context?: any) => {
+  logger.error('Application Error', {
+    message: error.message,
+    stack: error.stack,
+    context
+  });
+};
+
+// Função para log de ações de usuário
+export const logUserAction = (userId: string, action: string, details?: any) => {
+  logger.info('User Action', {
+    userId,
+    action,
+    details,
+    timestamp: new Date().toISOString()
+  });
+};
+
+// Substituir console.log em produção para manter logs estruturados
+if (process.env.NODE_ENV === 'production') {
+  console.log = (...args) => logger.info(args.join(' '));
+  console.error = (...args) => logger.error(args.join(' '));
+  console.warn = (...args) => logger.warn(args.join(' '));
+  console.info = (...args) => logger.info(args.join(' '));
+}
 
 export { logger };
 export default logger;
