--- backend/src/utils/logger.ts
@@ -1,101 +1,0 @@
-import winston from 'winston';
-import DailyRotateFile from 'winston-daily-rotate-file';
-import path from 'path';
-
-const logDir = process.env.LOG_DIR || 'logs';
-
-// Criar transporte para rotação diária de arquivos
-const fileRotateTransport = new DailyRotateFile({
-  filename: path.join(logDir, 'application-%DATE%.log'),
-  datePattern: 'YYYY-MM-DD',
-  maxSize: '20m',
-  maxFiles: '14d',
-  format: winston.format.combine(
-    winston.format.timestamp(),
-    winston.format.json()
-  )
-});
-
-// Criar transporte para erros
-const errorFileTransport = new DailyRotateFile({
-  filename: path.join(logDir, 'error-%DATE%.log'),
-  datePattern: 'YYYY-MM-DD',
-  level: 'error',
-  maxSize: '20m',
-  maxFiles: '30d',
-  format: winston.format.combine(
-    winston.format.timestamp(),
-    winston.format.json()
-  )
-});
-
-// Criar logger
-const logger = winston.createLogger({
-  level: process.env.LOG_LEVEL || 'info',
-  format: winston.format.combine(
-    winston.format.timestamp(),
-    winston.format.errors({ stack: true }),
-    winston.format.json()
-  ),
-  defaultMeta: { service: 'assist-move-marias' },
-  transports: [
-    fileRotateTransport,
-    errorFileTransport,
-    new winston.transports.Console({
-      format: winston.format.combine(
-        winston.format.colorize(),
-        winston.format.printf(({ timestamp, level, message, service, ...meta }) => {
-          return `${timestamp} [${service}] ${level}: ${message} ${Object.keys(meta).length ? JSON.stringify(meta) : ''}`;
-        })
-      )
-    })
-  ]
-});
-
-// Middleware para logs de requisições
-export const requestLogger = (req: any, res: any, next: any) => {
-  const startTime = Date.now();
-  
-  res.on('finish', () => {
-    const duration = Date.now() - startTime;
-    logger.info('HTTP Request', {
-      method: req.method,
-      url: req.url,
-      statusCode: res.statusCode,
-      duration: `${duration}ms`,
-      ip: req.ip,
-      userAgent: req.get('User-Agent')
-    });
-  });
-  
-  next();
-};
-
-// Função para log de erros com contexto
-export const logError = (error: Error, context?: any) => {
-  logger.error('Application Error', {
-    message: error.message,
-    stack: error.stack,
-    context
-  });
-};
-
-// Função para log de ações de usuário
-export const logUserAction = (userId: string, action: string, details?: any) => {
-  logger.info('User Action', {
-    userId,
-    action,
-    details,
-    timestamp: new Date().toISOString()
-  });
-};
-
-// Substituir console.log em produção para manter logs estruturados
-if (process.env.NODE_ENV === 'production') {
-  console.log = (...args) => logger.info(args.join(' '));
-  console.error = (...args) => logger.error(args.join(' '));
-  console.warn = (...args) => logger.warn(args.join(' '));
-  console.info = (...args) => logger.info(args.join(' '));
-}
-
-export default logger;
